moving the HUD to its own library
